tender_navigator_mvp/requirements.txt
streamlit
pypdf
python-docx
requests
pytest==8.3.5
